add error indication for node
