+++ node/vm/delete.go
@@ -6,11 +6,11 @@ import (
 	"github.com/yoneyan/vm_mgr/node/manage"
 )
 
-func DeleteVMProcess(id int) bool {
+func DeleteVMProcess(id int) (string, bool) {
 	result := manage.VMExistsID(id)
 	if result == false {
 		fmt.Println("VMID Not Found!!")
-		return false
+		return "VMID Not Found!!", false
 	}
 	err := VMStop(id)
 	if err != nil {
@@ -20,5 +20,5 @@ func DeleteVMProcess(id int) bool {
 	}
 
 	db.DeleteDBVM(id)
-	return true
+	return "ok", true
 }
